Task #5: Minor refactors

* Add VMAT T2

* Introduce machine specs.

* Make energy customizable.

* Merge branch 'feature/add_vmat_t2' into task/introduce-machine-classes

# Conflicts:
#	conjuror/plans/dicom.py

* Use machine classes in VMAT T2

* Annotate types.

* Revert "Annotate types."

This reverts commit 4511f5f35787602c6a2f5f3c1ea9eaaaef148836.

* Reapply "Annotate types."

This reverts commit d185837fac1da71bb1fb2b45eb0007963d39ec64.

* Doc strings

* Added input validation.

* Merge branch 'master' into task/doc-strings

# Conflicts:
#	conjuror/plans/dicom.py
#	tests/test_plan_generator.py


Approved-by: James Kerns

+++ conjuror/plans/fluence.py
@@ -18,7 +18,7 @@ def generate_fluences(
     ----------
     rt_plan : pydicom.Dataset
         The RT Plan dataset. Must contain BeamSequence.
-    width_mm : int
+    width_mm : float
         The width of the fluence map in mm. Use smaller values for faster calculation.
     resolution_mm : float, optional
         The resolution of the fluence map in mm. Smaller values will take longer to calculate.
